refactor gen package #117
