+++ src/main/java/cryptator/gen/member/CryptaMemberCard.java
@@ -6,7 +6,7 @@
  * Licensed under the BSD 3-clause license.
  * See LICENSE file in the project root for full license information.
  */
-package cryptator.gen;
+package cryptator.gen.member;
 
 import java.util.Arrays;
 import java.util.stream.IntStream;
@@ -14,6 +14,9 @@ import java.util.stream.IntStream;
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.constraints.extension.Tuples;
 import org.chocosolver.solver.variables.IntVar;
+
+import cryptator.gen.AbstractCryptaGenModel;
+import cryptator.gen.WordSumTuplesBuilder;
 
 public class CryptaMemberCard extends CryptaMemberLen {
