+++ src/main/java/cryptator/gen/pattern/CryptaLongMultModel.java
@@ -6,7 +6,7 @@
  * Licensed under the BSD 3-clause license.
  * See LICENSE file in the project root for full license information.
  */
-package cryptator.gen;
+package cryptator.gen.pattern;
 
 import java.util.Optional;
 import java.util.stream.IntStream;
